requirements.txt
torch>=2.2
transformers>=4.49
accelerate>=1.2
peft>=0.14
pillow
pandas
pyyaml
h5py
tqdm
torchmetrics>=1.4
wandb>=0.18
